Refactoring and changing name
refactoring the different name and changed the different classes, interface.
