--- HabitTrackerApp/app/src/main/java/com/example/habittrackerapp/model/ProfileData.kt
@@ -1,6 +1,0 @@
-package com.example.habittrackerapp.model
-
-data class ProfileData  (
-    var name: String = "",
-    var counter: Int = 0
-)
