+++ HabitTrackerApp/app/src/main/java/com/example/habittrackerapp/model/UserDataRepository.kt
@@ -1,0 +1,12 @@
+package com.example.habittrackerapp.model
+
+import com.example.habittrackerapp.user.User
+import kotlinx.coroutines.flow.Flow
+
+interface UserDataRepository {
+    suspend fun saveUser(profileData: User)
+
+    suspend fun getUser(): Flow<User>
+
+    suspend fun clear()
+}
